+++ manul/src/protocol.rs
@@ -33,4 +33,4 @@ pub use serialization::{Deserializer, Serializer};
 
 pub(crate) use errors::ReceiveErrorType;
 pub(crate) use message::ProtocolMessagePartHashable;
-pub(crate) use object_safe::{BoxedRng, ObjectSafeRound};
+pub(crate) use object_safe::ObjectSafeRound;